pyproject.toml
[build-system]
requires = [
    "setuptools",
    "wheel",
    "setuptools-git-versioning>=2.0,<3",
]
build-backend = "setuptools.build_meta"

[project]
name = "py-sam"
authors = [
    {name = "Lou Markovski", email = "lou.markovski@gmail.com"},
]
description = "Segment Anything Model 2 batch image processing with Python Ray"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = {file = "LICENSE.md"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "clip @ git+https://github.com/openai/CLIP.git@main",
    "httpx>=0.27.0",
    "huggingface-hub>=0.24.5",
    "matplotlib<3.10.0",  # See https://github.com/RVC-Project/Retrieval-based-Voice-Conversion-WebUI/issues/2411
    "numpy>=1.23.0",
    "sam-2 @ git+https://github.com/facebookresearch/sam2.git@main",
    "opencv-python-headless>=4.10.0.84",
    "pandas>=2.2.2",
    "pyarrow>=17.0.0",
    "python-dotenv>=1.0.1",
    "s3fs>=2024.6.1",
    "structlog>=24.2.0",
    "ray[default]==2.41.0",
    "tenacity>=8.4.2",
    "torch>=2.3.1",
    "typer>=0.12.3",
    "urllib3>=2.2.2",
]
dynamic = ["version"]

[tool.setuptools-git-versioning]
enabled = true
version_file = "src/py_sam/VERSION"

[project.optional-dependencies]
dev = [
    "black",
    "build",
    "deptry",
    "mkdocstrings-python",
    "mypy",
    "pytest",
    "pytest-cov",
    "pytest-sugar",
    "ruff",
    "structlog",
    "twine",
    "typer",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
pysam = "py_sam.__main__:app"

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
    # pylint
    "PLC", "PLE", "PLR", "PLW",
    # pep8-naming
    "N",
    # pydocstyle
    "D",
    # pylint
    "PLC", "PLE", "PLR", "PLW",
    # pep8-naming
    "N",
    # pydocstyle
    "D",
]
ignore = [
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
"__init__.py" = ["F401"]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.coverage.report]
skip_empty = true
show_missing = true
fail_under = 80

[tool.coverage.run]
branch = true
omit = [
    ".venv/*",
    "src/logging_config.py",
    "src/py_sam/__main__.py",
    "tests/*"
]

[tool.deptry.per_rule_ignores]
DEP001 = ["logging_config", "dotenv", "sam2"]
DEP002 = ["clip", "pandas"]
DEP003 = ["py_sam", "rich", "pyarrow"]

[tool.deptry]
exclude = [".pyenv",".venv"]
ignore = ["DEP004"]
